Custom keymap for V8 LE 7.0.2
This is a custom keymap for the V8 on LE 7.0.2. It is made using the Keymap Editor addon and the original tlbb keymap as a template. This keymap should require no additional files. If you've allready dropped in a custom remote.conf, that needs to be removed in order for this keymap to work. Drop this file in the ~/userdata/keymaps/-folder and reboot the box or do a "ReloadKeymaps".